Fix construction of authors array of hashes.
We want to be able to pass a Hash to `Package#save` as it was coming from a
standard form submission (so that we can rely on Rails conventions). Turns out
that if we want to `accept_nested_attributes` on Package we need to provide an
array containing the `Author`s hashes of attributes under the
`authors_attributes` key.